3 Main Issues
Hi All, My progress with this code has come to a halt. I've hit three
issues which I've outlined below. Any help / advice on where my code is
incorrect would be appreciated.

(1) I'm unable to get the callback function on the x-ray scraper add a
time key and value.

(At the moment the csv file isn't set up to take the modified data with
time from the scraper file, it is itself still taking the data directly
from a results.JSON file. For the time-being I've just been trying to
get the scraper to add the time to the returned scraped object before
later re-wiring the csv file accordingly).

(2) For the task 'If http://shirts4mike.com is down, an error message
describing the issue should appear in the console' I thought putting a
try/catch statement around the scraper function would do the trick,
however the only error I'm getting in the console/terminal is one from
events.js. I can't figure out how to capture this and make a
human-readable error message.

`events.js:160
      throw er; // Unhandled 'error' event
      ^

Error: getaddrinfo ENOTFOUND shirts4mike.com shirts4mike.com:80
    at errnoException (dns.js:28:10)
    at GetAddrInfoReqWrap.onlookup [as oncomplete] (dns.js:76:26)`

(3) I'm using moment.js to get the time. When I first run the code the
time is OK, but if I run the code a few minutes later, or half an hour
later, the time hasn't changed from the original. At a loss as to why
this would be.