3 Outstanding Questions
I've a couple of issues remaining that I've not been able to figure out
(1) I'm using moment.js to capture the data and time. The date is
always fine, but the time is always out by different amounts (It can be
out by anything from 1 minute to an hour). My system time seems fine,
it's just not matching up with moment.js. I've not seen anything on
Stack Overflow or the moment.js issues page about this. Wondering if
anyone else has had similar problems?

(2) Error messages.
I've put a try/catch block around the scraper request to catch an error
when shirts4mike is down. When I test with the wifi turned off a stack
trace is returned, but I can't access any specific error code on the
err object.

I've followed Andrew's 'Handling the Error Event in Node' vids and I
_think_ I've set it up the same way...The only difference I can see is
that I'm not using an http module therefore maybe the error object
doesn't come back the same?

Is it OK to simply write in a console.log message for when an
unspecified error occurs to fulfil this challenge's requirement? (as
one line 51). I think the challenge is looking for a different
approach, but I can't get to the specific errors in order to use.

(3) If I wanted to have a go at the Extra Credit challenge of logging
errors to a scraper-error.log file, would using a module such as
winston be acceptable?